APPIMAGE: Bundle $EXECUTABLE if passed.
Allows for pointing at arbitrary executable to bundle. Makes it easier
to experiment with static builds, and allows for a more generic CI
config.

+++ misc/appimage/appimage-manual_creation.sh
@@ -80,16 +80,19 @@ mkdir -p "$DIR/AppDir/usr/lib-override" || exit 1
 VERSION=$(sed -n 's/.*VERSION_NUMBER.*"\(.*\)".*/\1/p' src/version.h)
 REVISION=$(git log -n 1|head -1|awk '{print $2}'|cut -c1-6)
 
-#build ezquake
-export SKIP_DEPS
-chmod +x ./build-linux.sh && nice ./build-linux.sh || exit 3
+#build ezquake unless executable set
+if ! [[ -x "$EXECUTABLE" ]]; then
+	export SKIP_DEPS
+	chmod +x ./build-linux.sh && nice ./build-linux.sh || exit 3
+	EXECUTABLE="build/ezquake-linux-$ARCH"
+fi
 
 #build test program
 echo "$TESTPROGRAM" > "$DIR/test.c" || exit 2
 gcc "$DIR/test.c" -o "$DIR/AppDir/usr/bin/test" -lcurl || exit 2
 rm -f "$DIR/AppDir/test.c" || exit 2
 
-cp -f build/ezquake-linux-$ARCH "$DIR/AppDir/usr/bin/." || exit 4
+cp -f $EXECUTABLE "$DIR/AppDir/usr/bin/." || exit 4
 rm -f "$DIR/AppDir/AppRun"
 echo "$QUAKE_SCRIPT" > "$DIR/AppDir/AppRun" || exit 4
 chmod +x "$DIR/AppDir/AppRun" || exit 4
